Add development philosophy and fix Phase 1 GPU sign-conversion warnings
Introduce PHILOSOPHY.md mapping the "Reality 101" principles to Blended's
fork design, development workflow, and debugging methodology. The philosophy
drives concrete practices: fix warnings with simple casts not frameworks,
trust documented pitfalls, heal one subsystem at a time (fractal fixes),
and don't let recurring bugs gaslight you into wrong theories.

Integrate philosophy references into WARNINGS.md, README, UPSTREAM_SYNC.md,
and WEBASSEMBLY_ROADMAP.md so the principles are visible where decisions
are made.

Begin Phase 1 warning fixes in the GPU/WebGL2 backend:
- Fix uint16_t |= (1 << slot) sign-conversion warnings with explicit
  uint16_t(1) casts across 7 files (gl_storage_buffer, gl_uniform_buffer,
  gl_immediate, gl_index_buffer, gl_vertex_buffer)
- Fix uint8_t |= (1ULL << unit) narrowing in gl_state.cc image binds
- Fix GLenum + int sign-conversion in cube map face calculations
  (gl_texture.cc, gl_framebuffer.cc)
- Relax GL 4.3 version gate for Emscripten in gl_backend.cc (WebGL2 is
  GL ES 3.0; compute/SSBO gaps handled via separate fallback paths)

+++ source/blender/gpu/opengl/gl_immediate.cc
@@ -74,9 +74,9 @@ uchar *GLImmediate::begin()
 #ifndef NDEBUG
   if (this->shader->is_polyline) {
     /* Silence error. These are bound inside `immEnd()`. */
-    GLContext::get()->bound_ssbo_slots |= 1 << GPU_SSBO_POLYLINE_POS_BUF_SLOT;
-    GLContext::get()->bound_ssbo_slots |= 1 << GPU_SSBO_POLYLINE_COL_BUF_SLOT;
-    GLContext::get()->bound_ssbo_slots |= 1 << GPU_SSBO_INDEX_BUF_SLOT;
+    GLContext::get()->bound_ssbo_slots |= uint16_t(1) << GPU_SSBO_POLYLINE_POS_BUF_SLOT;
+    GLContext::get()->bound_ssbo_slots |= uint16_t(1) << GPU_SSBO_POLYLINE_COL_BUF_SLOT;
+    GLContext::get()->bound_ssbo_slots |= uint16_t(1) << GPU_SSBO_INDEX_BUF_SLOT;
   }
 #endif
 
@@ -165,9 +165,9 @@ void GLImmediate::end()
     this->polyline_draw_workaround(0);
 
 #ifndef NDEBUG
-    GLContext::get()->bound_ssbo_slots &= ~(1 << GPU_SSBO_POLYLINE_POS_BUF_SLOT);
-    GLContext::get()->bound_ssbo_slots &= ~(1 << GPU_SSBO_POLYLINE_COL_BUF_SLOT);
-    GLContext::get()->bound_ssbo_slots &= ~(1 << GPU_SSBO_INDEX_BUF_SLOT);
+    GLContext::get()->bound_ssbo_slots &= uint16_t(~(uint16_t(1) << GPU_SSBO_POLYLINE_POS_BUF_SLOT));
+    GLContext::get()->bound_ssbo_slots &= uint16_t(~(uint16_t(1) << GPU_SSBO_POLYLINE_COL_BUF_SLOT));
+    GLContext::get()->bound_ssbo_slots &= uint16_t(~(uint16_t(1) << GPU_SSBO_INDEX_BUF_SLOT));
 #endif
   }
   else {
