+++ source/blender/gpu/opengl/gl_vertex_buffer.cc
@@ -95,7 +95,7 @@ void GLVertBuf::bind_as_ssbo(uint binding)
 
 #ifndef NDEBUG
   BLI_assert(binding < 16);
-  GLContext::get()->bound_ssbo_slots |= 1 << binding;
+  GLContext::get()->bound_ssbo_slots |= uint16_t(1) << binding;
 #endif
 }
